// src/components/DmapglMap/MarkerCluster/MarkerCluster.types.ts
// Generated with util/create-component.js
export interface MarkerClusterProps {
    map?: any;
    id: string;
    data: any;
    markerFactory?:any,
    clusterFactory?:any,
    markerClick?:any,
    markerLeave?:any,
    markerEnter?:any,
    clusterMaxZoom?:number,
    clusterRadius?:number,
    useCluster?: boolean
  }